backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
python-dotenv
supabase
openai
faiss-cpu
sentence-transformers[onnx]
elevenlabs
edge-tts
